added search bar functionality
