+++ src/JS_Modules/Controllers/SearchBar.js
@@ -1,0 +1,13 @@
+import { getMovies } from "../Models/Services";
+import ListRender from "./ListRender";
+
+const searchBar = async (searchForm) => {
+    console.log(searchForm);
+    if (searchForm[0].value.trim().length > 0) {
+        const movieTitle = searchForm[0].value.trim();
+        const moviesFetch = await getMovies(movieTitle);
+        ListRender(moviesFetch);
+    }
+};
+
+export default searchBar;
