+++ src/JS_Modules/Views/Header.js
@@ -10,7 +10,7 @@ const header = () => `<header class="header-section">
         <div class="col-md-6 col-sm-12">
             <form method="post" class="searchMovie">
                 <div class="input-group">
-                        <input type="text" class="form-control search_movie" placeholder="Search a movie" aria-label="Search a movie" aria-describedby="search_movie" required />
+                        <input type="text" class="form-control search_movie" id="movieTitle" placeholder="Search a movie" aria-label="Search a movie" aria-describedby="search_movie" required />
                         <div class="input-group-append">
                         <button type="submit" class="input-group-text" id="search_movie"><i class="fa fa-film">&nbsp;</i>Search</button>
                         </div>
